Add dynamic facets to dataset search
