+++ src/main/java/io/github/genomicdatainfrastructure/discovery/filters/application/ports/FilterBuilder.java
@@ -10,5 +10,7 @@ import java.util.List;
 
 public interface FilterBuilder {
 
+    String source();
+
     List<Filter> build(String accessToken, String preferredLanguage);
 }
