+++ src/test/resources/mappings/package_search_datasets_when_query_ckan_and_beacon.json
@@ -90,7 +90,19 @@
                         "relationships_as_subject": [],
                         "relationships_as_object": []
                     }
-                ]
+                ],
+                "search_facets": {
+                    "tags": {
+                        "title": "Keywords",
+                        "items": [
+                            {
+                                "name": "synthetic",
+                                "display_name": "synthetic",
+                                "count": 1
+                            }
+                        ]
+                    }
+                }
             }
         }
     }
